refacor: ProfilePreviewStep 로직 분리
